Move core initialstate and mutations to own modules
